ADD: services

+++ src/main/java/com/gsc/programaavisos/repository/crm/PARepository.java
@@ -3,5 +3,5 @@ package com.gsc.programaavisos.repository.crm;
 import com.gsc.programaavisos.model.crm.entity.ProgramaAvisos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PARepository extends JpaRepository<ProgramaAvisos, Long>, PACustomRepository {
+public interface PARepository extends JpaRepository<ProgramaAvisos, Integer>, PACustomRepository {
 }
